move hex value parsing to separate library
