+++ shorthand/utils/rec.py
@@ -2,6 +2,7 @@ import re
 import json
 
 from shorthand.utils.record_set import RecordSet
+from shorthand.utils.rec_lib import get_hex_int
 
 
 ALLOWED_CONFIG_KEYS = [
@@ -19,25 +20,6 @@ PRIMITIVE_TYPES = [
 ALL_TYPES = [
     "int", "line", "date", "bool", "real", "uuid",
     "range", "enum", "size", "regexp"]
-
-
-def get_hex_int(hex_string):
-    '''Parse a hexadecimal string and return
-    the parsed value as an integer
-    '''
-
-    is_negative = False
-    unsigned_hex_string = hex_string
-
-    if hex_string[0] == '-':
-        is_negative = True
-        unsigned_hex_string = hex_string[1:]
-    if len(unsigned_hex_string) < 3:
-        raise ValueError('Invalid_hex_string')
-    int_value = int(unsigned_hex_string[2:], 16)
-    if is_negative:
-        int_value = int_value * -1
-    return int_value
 
 
 def load_from_file(file_path):
